Refactor map sheet
