+++ src/page/Map/index.tsx
@@ -4,17 +4,14 @@ import useUpdatePlaces from "@/hooks/api/place/useUpdatePlaces";
 import useBottomSheetSnapPoints from "@/hooks/useBottomSheetSnapPoints";
 import useCheckLoginAndRoute from "@/hooks/useCheckLoginAndRoute";
 import useMapSetup from "@/hooks/useMapSetup";
-import noReviews from "@/image/icons/receiptLines.svg";
 import {
   GetPlaceResponse,
   placeCategory,
   placeSort,
 } from "@/interface/apiInterface";
-import { H3 } from "@/style/font";
 import { getLastKnownPositionObj } from "@/utils/getFromLocalStorage";
 import useToastPopup from "@/utils/toastPopup";
-import { useCallback, useEffect, useRef, useState } from "react";
-import { Sheet, SheetRef } from "react-modal-sheet";
+import { useCallback, useEffect, useState } from "react";
 import {
   Container as MapDiv,
   NaverMap,
@@ -23,11 +20,8 @@ import {
 } from "react-naver-maps";
 import { useNavigate, useSearchParams } from "react-router-dom";
 import styled from "styled-components";
-import ReviewHeader from "./_components/headers/ReviewHeader";
-import SearchHeader from "./_components/headers/SearchHeader";
+import MapSheet from "./_components/index/MapSheet";
 import PinMarker from "./_components/PinMarker";
-import Restaurant from "./_components/Restaurant";
-import Review from "./_components/review/Review";
 import UserPositionMarker from "./_components/UserPositionMarker";
 
 const MapPage: React.FC = () => {
@@ -54,15 +48,6 @@ const MapPage: React.FC = () => {
   const defaultZoom = 20;
   const [isGeoAvailable, setIsGeoAvailable] = useState(false);
   useMapSetup(true, map, user, followUser, setKakaoPlaceId, setIsGeoAvailable);
-
-  console.log(
-    "id",
-    kakaoPlaceId,
-    "isreviewview",
-    isReviewView,
-    "follow",
-    followUser
-  );
 
   const { data: placeData } = useGetSpecificPlaces({
     kakaoPlaceId: kakaoPlaceId!,
@@ -112,26 +97,7 @@ const MapPage: React.FC = () => {
   }, [kakaoPlaceId]);
 
   // Bottom sheet logic
-  const sheetRef = useRef<SheetRef>();
-  const { snapPoints, attachRef, sheetHeaderRef } = useBottomSheetSnapPoints();
-  const [left, setLeft] = useState(0);
-  useEffect(() => {
-    const updateLeftPosition = () => {
-      const newLeft =
-        window.innerWidth > 440 ? (window.innerWidth - 440) / 2 : 0;
-      if (newLeft !== left) setLeft(newLeft);
-    };
-    updateLeftPosition();
-    window.addEventListener("resize", updateLeftPosition);
-    return () => {
-      window.removeEventListener("resize", updateLeftPosition);
-    };
-  }, [left]);
-
-  const removeQueries = () => {
-    const path = window.location.pathname; // 현재 경로
-    window.history.pushState({}, "", path); // 쿼리 없이 경로만 유지
-  };
+  const { attachRef } = useBottomSheetSnapPoints();
 
   // Track mouse down
   // Don't call places API while dragging
@@ -225,95 +191,20 @@ const MapPage: React.FC = () => {
                 />
               ))}
           </NaverMap>
-          <StSheet
-            ref={sheetRef}
-            isOpen={true}
-            onClose={() => {}}
-            snapPoints={snapPoints}
-            initialSnap={1}
-            mountPoint={attachRef.current!}
-            $left={left}
-          >
-            <Sheet.Container>
-              <Sheet.Header ref={sheetHeaderRef}>
-                <Sheet.Header />
-                {!isReviewView && (
-                  <SearchHeader
-                    dataQuery={dataQuery}
-                    setDataQuery={setDataQuery}
-                    setSort={setSort}
-                    category={category}
-                    setCategory={setCategory}
-                  />
-                )}
-                {isReviewView && (
-                  <ReviewHeader
-                    bookmark={bookmark}
-                    setBookmark={setBookmark}
-                    onBack={() => {
-                      removeQueries();
-                      setKakaoPlaceId(null);
-                    }}
-                  />
-                )}
-              </Sheet.Header>
-              <Sheet.Content style={{ paddingBottom: sheetRef.current?.y }}>
-                <Sheet.Scroller>
-                  {((!isReviewView && !places) ||
-                    (!isReviewView && places?.length === 0)) && (
-                    <div className="no-reviews">
-                      <img src={noReviews} />
-                      <p>근처에 리뷰 있는</p>
-                      <p>가게가 없어요!</p>
-                    </div>
-                  )}
-                  {places &&
-                    !isReviewView &&
-                    places.map((item) => (
-                      <div
-                        key={item.kakaoPlaceId}
-                        onClick={() => {
-                          setKakaoPlaceId(item.kakaoPlaceId);
-                          navigate(
-                            `${window.location.pathname}?kakaoPlaceId=${item.kakaoPlaceId}`
-                          );
-                        }}
-                      >
-                        <Restaurant
-                          key={item.kakaoPlaceId}
-                          name={item.name}
-                          averageStarRating={item.averageStarRating}
-                          reviewImageUrls={item.reviewImageUrls
-                            .slice()
-                            .reverse()}
-                          reviewerProfileImageUrls={item.reviewerProfileImageUrls
-                            .slice()
-                            .reverse()}
-                          reviewCount={item.reviewCount}
-                          distance={item.distance}
-                        />
-                      </div>
-                    ))}
-                  {isReviewView && (
-                    <Review
-                      setBookmark={setBookmark}
-                      currentLatitude={
-                        getLastKnownPositionObj().coords.latitude
-                      }
-                      currentLongitude={
-                        getLastKnownPositionObj().coords.longitude
-                      }
-                    />
-                  )}
-                </Sheet.Scroller>
-              </Sheet.Content>
-              <StGap
-                $attachHeight={
-                  attachRef.current?.offsetHeight ?? window.innerHeight - 65
-                }
-              />
-            </Sheet.Container>
-          </StSheet>
+          <MapSheet
+            {...{
+              sheet: { attachRef },
+              mapActions: {
+                setCategory,
+                setSort,
+                setDataQuery,
+                setBookmark,
+                setKakaoPlaceId,
+              },
+              mapData: { places },
+              mapState: { isReviewView, category, dataQuery, bookmark },
+            }}
+          />
         </StMapDiv>
       </NavermapsProvider>
     </StDiv>
@@ -337,52 +228,6 @@ const StButton = styled(Button)<{ $enabled: boolean }>`
 const StMapDiv = styled(MapDiv)`
   width: 100%;
   height: 100%;
-`;
-
-const StSheet = styled(Sheet)<{ $left: number }>`
-  display: flex;
-  justify-content: center;
-  max-width: var(--max_width);
-  min-width: var(--min_width);
-  left: ${({ $left }) => `${$left}px !important`};
-
-  .react-modal-sheet-content::-webkit-scrollbar {
-    width: 8px;
-  }
-
-  .react-modal-sheet-content::-webkit-scrollbar-track {
-    background: #f1f1f1;
-    border-radius: 4px;
-  }
-
-  .react-modal-sheet-content::-webkit-scrollbar-thumb {
-    background: #888;
-    border-radius: 4px;
-  }
-
-  .react-modal-sheet-content::-webkit-scrollbar-thumb:hover {
-    background: #555;
-  }
-
-  .no-reviews {
-    ${H3}
-    height:80%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    color: var(--neutral_500);
-    gap: var(--spacing_8);
-    margin-top: var(--spacing_12);
-
-    img {
-      height: 32px;
-    }
-  }
-`;
-
-const StGap = styled.div<{ $attachHeight: number }>`
-  height: ${({ $attachHeight }) => `${window.innerHeight - $attachHeight}px`};
 `;
 
 export default MapPage;
